Fix deny-all egress NetworkPolicy case (#573)
For egress policy, if there are no "ToAddresses" and "IPBlocks", there
were no PolicyRules created, which caused the Pods not being isolated.

We must ensure there is at least one PolicyRule, this patch creates a
PolicyRule with the original services if it doesn't exist. If there are
IPBlocks or Pods that cannot resolve any named port, they will share
this PolicyRule.