Handle DeletedFinalStateUnknown in NetworkPolicy Controller
According to the K8s documentation, DeletedFinalStateUnknown is used
when an object was deleted but the watch deletion event was missed. In
my testing, I have observed this situation with Antrea running in Kind,
when the Antrea Controller loses connectivity to the K8s apiserver
because the Agent running on the same Node is restarted.

The code is adapted from the K8s sample controller and from K8s control
plane controllers.

Fixes #579