Use correct URL for Windows OVS binaries
At the moment the archive can only be downloaded over HTTP and not
HTTPS (because of how virtual hosting works with AWS S3). We are working
on getting HTTPS downloads to work for downloads.antrea.io. In the
meantime we verify that the SHA256 hash for the downloaded archive is
correct.