Fix NetworkPolicy failures in Kind clusters
When running NetworkPolicy tests in Kind, we "force" the tunnel type to
VXLAN as we observe some failures with Geneve (see #897).

We also change the Github CI jobs so that CNP tests are run in their own
job. This is to address the issue above, and is also a partial
implementation of #893.