Revert "klogv2 upgrade and klogv1 redirect (#966)" (#1083)
This reverts commit 211852902046d004b05a74a177ec06285e8426e7.

The code to redirect klogv1 to klogv2 (the version of the K8s libraries
we use - 0.18.4 - still use klogv1) is incorrect, leading to truncated
log messages, or log messages with incorrect source file information.

Since the code to fix it is pretty complex (see
https://github.com/vmware-tanzu/antrea/pull/1082) and there is no
apparent benefit to using klogv2 in Antrea, we are just reverting the
commit which switched from klogv1 to klogv2. When we update to a more
recent version of the k8s Golang libraries, we will transition Antrea to
klogv2 as well.

Note that one dependency (k8s.io/utils) was already using klogv2. We
rolled back the version of that dependency by a couple commits in go.mod
to ensure that none of the software is using klogv2.

Fixes #1075