Controller support for ExternalEntity in Antrea NetworkPolicy (#1084)
* Agent supports ExternalEntity.

This commit moves ToAddresses/FromAddresses in CompletedRule and AddressSetByGroup in ruleCache
to use GroupMemberSet instead of GroupMmeberPodSet. Thus both Pods and ExternalEntities are
expressed as GroupMember when in these fields.

Pods in appliedTo field continue be expressed by existing GroupMemberPod, and migration to GroupMember shall
be done in a subsequent PR.

* Add support for ANP and externalEntities in controller

* Unify functions for Pod and ExternalEntity in networkpolicy controller

* Improve UT coverage

* Address comments

* Resolve more comments

Co-authored-by: Su Wang <suw@vmware.com>